Settings_Extrapolation: Moved settings normalization tests into its own spec file
